# docs/swagger.yaml
basePath: /
definitions:
...
  models.AplicativoInfo:
    properties:
      device_id:
        type: string
      mac:
        type: string
      nome_do_aplicativo:
        type: string
      vencimento_aplicativo:
        type: string
    type: object
...
  models.EditUserRequest:
    properties:
      aplicativos:
        items:
          $ref: '#/definitions/models.AplicativoInfo'
        type: array
      bouquet:
        type: string
      enviar_notificacao:
        type: boolean
      nome_para_aviso:
        type: string
      numero_whats:
        type: string
      password:
        type: string
      reseller_notes:
        type: string
      username:
        type: string
    type: object
  models.ScreenRequest:
    properties:
      userID:
        type: integer
    required:
    - userID
    type: object
...
host: localhost:8080
info:
  contact:
    email: suporte@example.com
    name: Suporte
  description: Documentação da API IPTV
  title: API IPTV
  version: 1.0.5
paths:
...
  /api/tools-table/add-screen:
    post:
      consumes:
      - application/json
      description: Aumenta o número máximo de conexões do usuário e desconta créditos
        se aplicável
      parameters:
      - description: JSON contendo o ID do usuário
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ScreenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Retorna o novo total de telas e o saldo de créditos atualizado
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Erro nos parâmetros ou créditos insuficientes
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno ao adicionar tela
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Adiciona uma nova tela ao usuário
      tags:
      - ToolsTable
  /api/tools-table/edit/{id}:
    put:
      consumes:
      - application/json
      description: Permite a edição de dados de um usuário na revenda autenticada
      parameters:
      - description: ID do usuário a ser editado
        in: path
        name: id
        required: true
        type: integer
      - description: Dados do usuário a serem editados
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.EditUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Usuário editado com sucesso
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Erro na requisição
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido ou acesso negado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno ao processar a requisição
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Edita um usuário
      tags:
      - ToolsTable
  /api/tools-table/remove-screen:
    post:
      consumes:
      - application/json
      description: Diminui o número máximo de conexões do usuário, garantindo que
        tenha pelo menos uma tela ativa
      parameters:
      - description: JSON contendo o ID do usuário
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ScreenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Retorna o novo total de telas
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Erro nos parâmetros ou limite mínimo atingido
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno ao remover tela
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove uma tela do usuário
      tags:
      - ToolsTable
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"